refactor(frontend): :fire: Remove unused code, sort and organize files
