+++ frontend/.prettierignore
@@ -3,4 +3,4 @@ build
 coverage
 
 # Ignore gql.tada generated types
-src/tada/graphql-env.d.ts
+src/lib/tada/graphql-env.d.ts
