+++ frontend/src/app/_homeComponents/FeaturedProjectCard.tsx
