+++ frontend/src/app/_homeComponents/FeaturedProjects.tsx
@@ -1,7 +1,7 @@
 import fetchFeaturedProjects from "@/lib/data/fetchFeaturedProjects";
 import Link from "next/link";
-import { Button } from "./shadcn/button";
-import CardWithContent from "./CardWithContent";
+import { Button } from "../shadcn/button";
+import CardWithContent from "../shared/CardWithContent";
 
 const STRAPI_URL = process.env.BACKEND_INTERNAL_URL;
