+++ frontend/src/app/_homeComponents/MissionStatement.tsx
