+++ frontend/src/app/_homeComponents/SatelliteGlobe.tsx
