+++ frontend/src/app/_homeComponents/SatelliteSelector.tsx
