+++ frontend/src/app/_homeComponents/TeamSection.tsx
