+++ frontend/src/app/_homeComponents/files/globe-cables-data.json
