+++ frontend/src/app/_homeComponents/files/globe-data-min.json
