+++ frontend/src/app/_homeComponents/files/globe-data.json
