+++ frontend/src/app/blog/BlogpageButtons.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useState } from "react";
-import { Button } from "./shadcn/button";
+import { Button } from "../../components/shadcn/button";
 import { useRouter, useSearchParams } from "next/navigation";
 import { cn } from "@/lib/utils";
