+++ frontend/src/app/satellites/[satelliteSlug]/_2dmap/Map2d.tsx
