+++ frontend/src/components/layout/ErrorBoundaryNavigation.tsx
