+++ frontend/src/components/layout/Navbar.tsx
