+++ frontend/src/components/layout/Starfield.tsx
