+++ frontend/src/components/shared/BlockRendererClient.tsx
