+++ frontend/src/components/shared/CardWithContent.tsx
