+++ frontend/src/components/shared/RelatedProjectsAndSatellites.tsx
