+++ frontend/src/lib/SlicePreviewText.ts
