--- frontend/src/lib/data/fetchArticleInfo.tsx
@@ -1,93 +1,0 @@
-import { getClient } from "../ApolloClient";
-import { BlocksContent } from "@strapi/blocks-react-renderer";
-import { BlogPost } from "@/app/blog/page";
-import { graphql } from "@/tada/graphql";
-
-
-
-export default async function fetchArticlePages({
-    currentPage,
-    pageSize,
-    tag,
-}: {
-    currentPage: number;
-    pageSize: number;
-    tag: string | null;
-}) {
-    let firstArticle = true;
-
-    let graphqlData;
-
-    if (!tag) {
-        graphqlData = await getClient().query({
-            query: GET_ARTICLES,
-            variables: {
-                pagination: {
-                    pageSize: pageSize,
-                    page: currentPage,
-                },
-            },
-        });
-    } else {
-        
-    }
-    if (
-        graphqlData.data === null ||
-        graphqlData.data === undefined ||
-        graphqlData.data.articles === undefined ||
-        graphqlData.data.articles === null
-    ) {
-        return null;
-    }
-    let articleList: BlogPost[] = [];
-
-    graphqlData.data.articles.data.forEach((article: any) => {
-        let avatarURL: string | undefined =
-            article?.attributes?.author?.data?.attributes?.avatar?.data
-                ?.attributes?.url;
-
-        if (STRAPI_URL && avatarURL != undefined) {
-            avatarURL = STRAPI_URL + avatarURL;
-        }
-
-        const authorName: string | undefined =
-            article?.attributes?.author?.data?.attributes?.name;
-        const tag: string | null | undefined = article?.attributes?.Tag;
-        const datePublished: any = article?.attributes?.datePublished;
-        let coverImage: string | undefined =
-            article?.attributes?.coverImage?.data?.attributes?.url;
-        if (STRAPI_URL && coverImage != undefined) {
-            coverImage = STRAPI_URL + coverImage;
-        }
-        let content: BlocksContent = article?.attributes?.body ?? [];
-
-        const title: string | undefined = article?.attributes?.previewTitle;
-
-        for (const block of content) {
-            if (block.type === "paragraph") {
-                content = [block];
-                break;
-            }
-        }
-
-        articleList.push({
-            key: article.id,
-            firstArticle: firstArticle,
-            title,
-            content,
-            coverImage,
-            datePublished,
-            tag,
-            HOST_URL: STRAPI_URL,
-            authorName,
-            avatarURL,
-            slug: article.attributes.slug,
-        });
-
-        firstArticle = false;
-    });
-
-    const totalArticles = graphqlData.data?.articles?.meta?.pagination?.total;
-
-    return { articleList, totalArticles };
-}
